Improved `Yuni::Variant`, added support for arrays and method calls, plus various minor fixes
git-svn-id: http://svn.ta3d.org/yuni/trunk@2128 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/stream.hxx
@@ -287,7 +287,7 @@ namespace File
 		// Making sure that the buffer is zero-terminated if required
 		if (buffer.zeroTerminated)
 			*((C*)(buffer.data() + buffer.size())) = C();
-		return result;
+		return (uint) result;
 	}
 
 
@@ -310,7 +310,7 @@ namespace File
 		// Making sure that the buffer is zero-terminated if required
 		if (buffer.zeroTerminated)
 			*((C*)(buffer.data() + buffer.size())) = C();
-		return result;
+		return (uint) result;
 	}
